Report QUnit assertions as TestResults and map errors
First this adds an identifier to errors that we collect on QUnit's
testEnd event. This allows us to later remap these errors to assertions.
The reason we need to do this is because QUnit removes the expected and
actual props by the time the runEnd event fires.

This also treats each assertion as a TestResult.

These two things together get us much better test reporting output.